+++ src/Authentication/Repository/UserOauthRepository.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Authentication\Repository;
+
+use App\Authentication\Entity\UserOauth;
+use Doctrine\ORM\EntityRepository;
+
+/**
+ * @extends EntityRepository<UserOauth>
+ */
+class UserOauthRepository extends EntityRepository
+{
+}